+++ app/routes/searchRoomCommercial.tsx
@@ -1,8 +1,17 @@
 import { Card, CardGroup, CardHeader, Header, Label, Link, Select } from "@trussworks/react-uswds";
+import Breadcrumbs from "~/components/Breadcrumbs";
+import { mergeMeta } from "~/lib/merge-meta";
+
+export const meta = mergeMeta(({ parentTitle }) => [{ title: `Search for a room - Commercial • ${parentTitle}` }]);
 
 
-export default function searchRoomNonCommercial() {
-    
+export default function searchRoomCommercial() {
+    const breadcrumbLinks = [
+        { href: "#", text: "Home" },
+        { href: "#", text: "Meeting Spaces" },
+        { href: "#", text: "Search for a room - Commercial" },
+    ];
+
     return (
 
         <div className="flex justify-center">
@@ -19,17 +28,19 @@ export default function searchRoomNonCommercial() {
                         </div>
                     </Header>
                     <div className="ml-5 mr-5">
-
+                        <div className="-mt-1 ml-1">
+                            <Breadcrumbs links={breadcrumbLinks} />
+                        </div>
                         <CardHeader className="-mt-3">
-                            
-                            <br/>
                             <h1 className="font-sans text-[40px] usa-card__heading font-bold py-2">
                                 FIND A ROOM
                             </h1>
+                            <h1 className="usa-card__heading">COMMERCIAL/BUSINESS</h1>
+                            <br/>
 
                             <p className="font-sans text-base-darker text-sans-xs">
-                            For smaller groups, please consider our Shared Learning Rooms with capacities of 4, 8, and 10. 
-                            Groups up to 20+ to 100 people are better for Meeting Rooms.
+                            <strong className="font-bold">Rooms for commercial work purposes are only available at capacities of 4, 8, and 10. </strong>
+                            Rooms can be booked in 15 min time blocks up to 2-hours, and up to 2 weeks out.
                             </p>
                         </CardHeader>
                     </div>
